requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Neo4j Driver
neo4j==5.15.0

# Environment Variables
python-dotenv==1.0.0

# Streamlit
streamlit==1.31.0
streamlit-agraph==0.0.6
plotly==5.18.0

# AI/ML
google-generativeai==0.3.2

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.1.4
networkx==3.2.1

# HTTP Client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2

